+++ error.php
@@ -11,6 +11,7 @@ defined('_JEXEC') or die;
 use Joomla\CMS\Factory;
 use Joomla\CMS\HTML\HTMLHelper;
 use Joomla\CMS\Language\Text;
+use Joomla\CMS\Layout\FileLayout;
 use Joomla\CMS\Uri\Uri;
 
 /** @var Joomla\CMS\Document\HtmlDocument $this */
@@ -23,6 +24,9 @@ $pageclass     = $menu !== null ? $menu->getParams()->get('pageclass_sfx', '') :
 $themeSwitcher = (boolean)$this->params->get('theme-switcher', 1);
 $fontAwesome   = (boolean)$this->params->get('font-awesome-thats-actually-rather-shit', 1);
 $googleFont    = $this->params->get('google-font', '');
+$faviconPath   = (boolean)$this->params->get('favicon-override', 0)
+                 ? JPATH_ROOT
+                 : 'templates/' . $this->template . '/favicon';
 
 // Template params
 if ($themeSwitcher)
@@ -73,6 +77,12 @@ elseif ($this->params->get('siteTitle'))
 }
 
 $this->setMetaData('viewport', 'width=device-width, initial-scale=1');
+$this->setMetaData('theme-color', '#ffffff');
+
+$imageLayout = new FileLayout('lightning.favicon');
+$imageLayout->render([
+	'path' => 'templates/' . $this->template . '/favicon',
+]);
 ?>
 <!DOCTYPE html>
 <html lang="<?php echo $this->language; ?>" dir="<?php echo $this->direction; ?>">